a11y: Print script module live regions in page HTML (#65380)
Fix issues where @wordpress/a11y script module `speak` may not be
detected and announced by accessibility software, in particular when
dynamically loading the module.

Accessibility software needs to detect the live regions in order to
announce them. The `@wordpress/a11y` script module was adding the live
regions to the DOM correctly when the module was loaded, but when loaded
dynamically (via `import( '@wordpress/a11y' ).then( /*…*/ )` ) the live
region changes would happen before the live region was detected by
accessibility software and the announcements would not happen.

The `@wordpress/a11y` script module is updated so that it does not
inject the live regions into the DOM but expects them to be present on
the page.

A hook is added to inject the necessary HTML for the live regions when
the @wordpress/a11y script module may be present (enqueued or in the
import map).

This has a few additional advantages:

- Accessibility software correctly detects the live regions.
- The DOM does not need to be modified immediately on page load.
- The (gzipped) size of the script module is reduced by 416B (~46%).

The `wp-a11y` script shares the same DOM elements. It includes
protection to prevent injecting duplicate elements.

---

Co-authored-by: sirreal <jonsurrell@git.wordpress.org>
Co-authored-by: gziolo <gziolo@git.wordpress.org>
Co-authored-by: afercia <afercia@git.wordpress.org>

Source: https://github.com/WordPress/gutenberg/commit/05660fe9b2abf4e0f1cd74e9821cb7742772e327