Editor: Multi-entity saving: Show correct count of entities to be saved (#66482)
The modal was incorrectly displaying the number of different *types* of
entities to be saved. This result was inconsistent with the actual list
of entities rendered underneath and with the "Review _n_ changes" label
on the left-hand sidebar.

Source: https://github.com/WordPress/gutenberg/commit/0466f7d510ef5185d054b3d30b19f21ed9ad986b