Terms: Respect order specified by register_taxonomy() (#67154)
It is possible to supply a default set of `args` to `register_taxonomy()` which will be used when querying a list of terms -- for example, `orderby` in order to specify how the resulting list of terms should be sorted. This commit makes it so that the list of terms returned by the Terms REST API controller respects that order.

Co-authored-by: ockham <bernhard-reiter@git.wordpress.org>
Co-authored-by: jsnajdr <jsnajdr@git.wordpress.org>

Source: https://github.com/WordPress/gutenberg/commit/a6a04bebfb96f47df6c1f238c0fe63233d6dd3f5