Combine the release steps to ensure that releases are tagged (#65591)
Currently the release process is timing out. This is due to the 5 minute timeout included in the trunk commit command.

The current steps appear to be attempting to set the `Stable Tag` to the current stable (not the current release) but is setting it to the upcoming release anyway.

The process can be simplified by combining the Stable Tag / Trunk Commit / SVN copy into a single commit.

This will avoid situations where the trunk commit passes but the tag is not created.

Raising the timeout to 10 minutes (which is the default in SVN) should hopefully resolve the underlying timeout.

Co-authored-by: dd32 <dd32@git.wordpress.org>
Co-authored-by: andrewserong <andrewserong@git.wordpress.org>
Co-authored-by: mikachan <mikachan@git.wordpress.org>
Co-authored-by: ockham <bernhard-reiter@git.wordpress.org>
Co-authored-by: talldan <talldanwp@git.wordpress.org>

Source: https://github.com/WordPress/gutenberg/commit/2e34bedac3a35c58f5041e8add30566aac033c1d