Avoid using `*-latest` for runner images (#71201)
* Avoid using `*-latest` for runner images

While using the `ubuntu-latest`, `macos-latest`, and `windows-latest` runner image tags is convenient, it has proven to be problematic in a number of instances as the runners are slowly updated (see Core-62808 and Core-62843).

Additionally, GitHub uses warnings on workflow runs using the `-latest` tags when there are upcoming changes to the images that the alias points to. There is no way to acknowledge and dismiss these notices, and they can be quite noisy and alarming.

This switches all workflows to using specific version tags representing the latest non-preview versions, which currently are as follows:

- `ubuntu-24.04`
- `windows-2025`
- `macos-15`

* Only reference OS type in job name

This makes helps avoid needing to change the required check every time the runner is updated.

* Use `matrix`, not `inputs` as the context

---------

Co-authored-by: desrosj <desrosj@git.wordpress.org>
Co-authored-by: johnbillion <johnbillion@git.wordpress.org>

Source: https://github.com/WordPress/gutenberg/commit/82768d0e018b44007e3e3a25636c13375b33e7e7