Fix Navigation Link Focus Loss (#72071)
The setEditingControls state was causing the RichText element to steal focus when it was remounted when setEditingControls became false. The setEditingControls workaround was intended to fix this, but it appears the underlying issue that caused setEditingControls to be added is no longer an issue. Removing setEditingControls fixes the bug.

---------
Co-authored-by: jeryj <jeryj@git.wordpress.org>
Co-authored-by: getdave <get_dave@git.wordpress.org>

Source: https://github.com/WordPress/gutenberg/commit/710e544028730fd310078a90b6978988a2825f5f