RichText: Don't forget ref passing in Preview mode (#73062)
Let RichText forward the ref that it receives to the element it returns,
even when this element is the lightweight alternative returned when the
block editor is in preview mode.

Fixes any kind of bug arising from the fact that refs passed to RichText
were ignored when the block editor is in preview mode. Particularly,
`useBlockProps`, with its many merged refs, was systematically ignored.

This was revealed while debugging the bug that #73020 sought to fix:
`useBlockElement` could never return elements because the ref callback
of `useBlockElementRef` was never called.

Co-authored-by: mcsf <mcsf@git.wordpress.org>
Co-authored-by: jorgefilipecosta <jorgefilipecosta@git.wordpress.org>

Source: https://github.com/WordPress/gutenberg/commit/70f48bfe018d393b738bc3e3c88046378c775abe