Fix empty url value from unbinding entity from inspector sidebar (#72447)
When unsyncing a navigation link url, we now set the url as the synced entity url to avoid empty url states which cause bugs in the canvas. We also defer updating the url value in the store until the blur event to avoid the canvas receiving a temporarily empty value (such as when a user clears the URL in order to start typing a new URL).

----

Co-authored-by: jeryj <jeryj@git.wordpress.org>
Co-authored-by: getdave <get_dave@git.wordpress.org>
Co-authored-by: ockham <bernhard-reiter@git.wordpress.org>

Source: https://github.com/WordPress/gutenberg/commit/6a7bf43134230b854fe697b6dc275a659e2a9db0